SITL: make airspeed pin float

+++ libraries/AP_HAL_SITL/SITL_State.h
@@ -50,7 +50,7 @@ public:
 
     // simulated airspeed, sonar and battery monitor
     uint16_t sonar_pin_value;     // pin 0
-    uint16_t airspeed_pin_value;  // pin 1
+    float airspeed_pin_value;  // pin 1
     float voltage_pin_value;   // pin 13
     float current_pin_value;   // pin 12
 
@@ -129,7 +129,7 @@ private:
     void _output_to_flightgear(void);
     void _simulator_servos(SITL::Aircraft::sitl_input &input);
     void _simulator_output(bool synthetic_clock_mode);
-    uint16_t _airspeed_sensor(float airspeed);
+    float _airspeed_sensor(float airspeed);
     uint16_t _ground_sonar();
     float _rand_float(void);
     Vector3f _rand_vec3f(void);
